Sprint 3 - 20240523

+++ app/Http/Controllers/PaymentController.php
@@ -14,6 +14,7 @@ use Dompdf\Options;
 // Mailling
 use Illuminate\Support\Facades\Mail;
 use App\Mail\OrderCompleted;
+use Exception;
 
 class PaymentController extends Controller
 {
@@ -25,6 +26,7 @@ class PaymentController extends Controller
 
         switch ($request->payment_method) {
             case "paypal":
+                $order = $this->saveStatusOrder($request->orderId, -1, "paypal");
                 $provider = new PayPalClient();
                 $provider->setApiCredentials(config('paypal'));
                 $paypalToken = $provider->getAccessToken();
@@ -60,17 +62,21 @@ class PaymentController extends Controller
                 }
                 break;
             case "wire":
-                $order = Order::where("reference", 'LIKE', $request->orderId)->first();
-                $order->payment_method = "wire";
-                $order->save();
-                //
-                $this->generateOrderPdf($order);
-                Mail::to($order->email)->send(new OrderCompleted($order));
+                $order= $this->saveStatusOrder($request->orderId, 2,"wire");
+
+                $iban = \App\Http\Controllers\GeneralSettingController::getSetting('iban')->value;
+                if (!$iban) {
+                    return redirect()->route('payment.cancel', ['orderId' => $request->orderId])->with('error', __("This payment method is currently unavailable. Please choose another one."));
+                }
+                // dd($iban);
+
+                $this->generateOrderPdf($order, $iban);
+                Mail::to($order->email)->send(new OrderCompleted($order, $iban));
                 Cart::instance("default")->destroy();
-                return view('public.purchaseCompleted', compact('order'));
+                return view('public.purchaseCompleted', compact('order', 'iban'));
                 break;
             default:
-                $message = "Payment method not lp.\nPlease select another payment method";
+                $message = "Payment method not implemented.\nPlease select another payment method";
                 return redirect()->back()->with('error', $message);
                 break;
         }
@@ -120,79 +126,86 @@ class PaymentController extends Controller
     function success(Request $request)
     {
 
-        $provider = new PayPalClient();
-        $provider->setApiCredentials(config('paypal'));
-        $paypalToken = $provider->getAccessToken();
-        $provider->setCurrency('EUR');
-        $response = $provider->capturePaymentOrder($request->token);
+        try {
+            $provider = new PayPalClient();
+            $provider->setApiCredentials(config('paypal'));
+            $paypalToken = $provider->getAccessToken();
+            $provider->setCurrency('EUR');
+            $response = $provider->capturePaymentOrder($request->token);
 
-        if (isset($response['status']) && $response['status'] == "COMPLETED") {
-            $order = Order::where("reference", 'LIKE', $request->orderId)->first();
-            $order->status_id = 2;
-            $order->payment_method = "paypal";
-            $order->save();
+            if (isset($response['status']) && $response['status'] == "COMPLETED") {
+                $order = $this->saveStatusOrder($request->orderId, 3);
 
-            $orderStatusHistory = new OrderStatusHistory();
-            $orderStatusHistory->order_id = $order->id;
-            $orderStatusHistory->status_id = $order->status_id;
-            $orderStatusHistory->save();
-            //
-            $this->generateOrderPdf($order);
-            Mail::to($order->email)->send(new OrderCompleted($order));
-            Cart::instance("default")->destroy();
-            //dd( $order->details->first()->book->title);
-            return view('public.purchaseCompleted', compact('order'));
-        } else {
-            return redirect()->route('payment.cancel', ['orderId' => $request->orderId]);
+                $this->generateOrderPdf($order);
+                Mail::to($order->email)->send(new OrderCompleted($order));
+                Cart::instance("default")->destroy();
+
+                return view('public.purchaseCompleted', compact('order'));
+            } else {
+                return redirect()->route('payment.cancel', ['orderId' => $request->orderId]);
+            }
+        } catch (Exception $e) {
+            return redirect()->route('payment.cancel', ['orderId' => $request->orderId])->with('error', $e->getMessage());
         }
-        // dd($response);
     }
 
     function cancel(Request $request)
     {
-        $order = Order::where("reference", 'LIKE', $request->orderId)->first();
-        $order->status_id = 6;
-        $order->save();
-
-        $orderStatusHistory = new OrderStatusHistory();
-        $orderStatusHistory->order_id = $order->id;
-        $orderStatusHistory->status_id = $order->status_id;
-        $orderStatusHistory->save();
-
-        $orderId = $request->orderId;
-
-        // return redirect()->route('checkout.payment_method', ['orderId' => $request->orderId])->with(['error' => 'Error en la compra', 'shipment_options']);
-        // return view("public.payment", compact('orderId', 'shipment_options', 'shipment_tax'));
+        $this->saveStatusOrder($request->orderId, 6);
         return redirect()->route('checkout.toPaymentFromCancelled', ['orderId' => $request->orderId])->with('error', __('checkout.payment-error'));
     }
 
-    public function generateOrderPdf($order)
+    public function generateOrderPdf($order, $iban = null)
     {
-        // Obtener la orden y sus detalles
+        try {
+            // Obtener la orden y sus detalles
 
-        // Crear una nueva instancia de Dompdf
-        $pdf = new Dompdf();
-        $options = new Options();
-        $options->set('isHtml5ParserEnabled', true);
-        $pdf->setOptions($options);
+            // Crear una nueva instancia de Dompdf
+            $pdf = new Dompdf();
+            $options = new Options();
+            $options->set('isHtml5ParserEnabled', true);
+            $pdf->setOptions($options);
 
-        // HTML para el ticket de la orden
-        $html = view('public.order_pdf', compact('order'))->render();
+            // HTML para el ticket de la orden
+            $html = view('public.order_pdf', compact('order', 'iban'))->render();
 
-        // Cargar HTML en Dompdf
-        $pdf->loadHtml($html);
+            // Cargar HTML en Dompdf
+            $pdf->loadHtml($html);
 
-        // Renderizar PDF
-        $pdf->render();
+            // Renderizar PDF
+            $pdf->render();
 
 
-        // Guardar el PDF en una ubicación específica
-        $pdfFilePath = public_path('files/orders/' . $order->reference . '.pdf');
-        file_put_contents($pdfFilePath, $pdf->output());
+            // Guardar el PDF en una ubicación específica
+            $pdfFilePath = public_path('files/orders/' . $order->reference . '.pdf');
+            file_put_contents($pdfFilePath, $pdf->output());
 
-        // Devolver la ruta del archivo guardado
-        return $pdfFilePath;
-        // Descargar el PDF
-        //return $pdf->stream('order.pdf');
+            // Devolver la ruta del archivo guardado
+            return $pdfFilePath;
+            // Descargar el PDF
+            //return $pdf->stream('order.pdf');
+        } catch (Exception $e) {
+            return redirect()->route('checkout.toPaymentFromCancelled', ['orderId' => $order->id])->with('error', __('errors.unknown-error'));
+        }
+    }
+
+    private function saveStatusOrder($reference, $statusId = -1, $paymentMethod = null)
+    {
+        $order = Order::where("reference", 'LIKE', $reference)->first();
+
+        if ($paymentMethod != null) {
+            $order->payment_method = $paymentMethod;
+        }
+        if ($statusId != -1) {
+            $order->status_id = $statusId;
+            $orderStatusHistory = new OrderStatusHistory();
+            $orderStatusHistory->order_id = $order->id;
+            $orderStatusHistory->status_id = $order->status_id;
+            $orderStatusHistory->save();
+        }
+        $order->save();
+
+
+        return $order;
     }
 }
